Added TV show info view
When a guest or authenticated user clicks on a TV show name on the main page, it brings the user to a new info view, which displays all the info on the TV show, including actors, score, user's personalized rating and whether the show is added to the user's watchlist.

Everything is made to match the style of the main home page.